+++ next/src/App.tsx
@@ -6,7 +6,7 @@ import i18next from '@/lib/i18n'
 
 import ErrorComponent from './components/layout/ErrorComponent'
 import Layout from './components/layout/Layout'
-import DefaultPendingComponent from './components/ui/Waiting'
+import WaitingComponent from './components/layout/WaitingComponent'
 import { routeTree } from './routeTree.gen'
 
 const queryClient = new QueryClient()
@@ -27,7 +27,7 @@ const router = createRouter({
       />
     </Layout>
   ),
-  defaultPendingComponent: DefaultPendingComponent,
+  defaultPendingComponent: WaitingComponent,
 })
 
 // Register the router instance for type safety