+++ next/src/components/ui/Waiting.tsx
@@ -1,6 +1,0 @@
-import { useTranslation } from 'react-i18next'
-
-export default function DefaultPendingComponent() {
-  const { t } = useTranslation()
-  return <>{t('common.loading')}</>
-}