+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Get list of variables scopes by calling the backend
 - Display scope label in variables Page
+- Default pending component (to replace by Skeleton)
 
 ## [3.2.3](https://github.com/InseeFr/Pogues/releases/tag/3.2.3) - 2026-06-04